Moved get_account_history from database_api to account_history. #573
